Remove P2PHelpers and simplify constructors in MsbuildProject
